separated controllers from auth routes
