+++ server/controllers/authController.js
@@ -1,0 +1,37 @@
+
+const User = require('../models/User');
+const jwt = require('jsonwebtoken');
+require('dotenv').config();
+
+exports.registerUser = async (req, res) => {
+    const { username, email, password } = req.body;
+    try {
+        const userExists = await User.findOne({ email });
+        if (userExists) {
+            return res.status(400).json({ message: 'User already exists' });
+        }
+
+        const user = new User({ username, email, password });
+        await user.save();
+        res.status(201).send('User registered successfully');
+    } catch (error) {
+        res.status(500).json({ message: error.message });
+    }
+};
+
+exports.loginUser = async (req, res) => {
+    const { email, password } = req.body;
+    try {
+        const user = await User.findOne({ email });
+        if (!user) return res.status(400).json({ message: 'Invalid email or password' });
+
+        const isMatch = await user.comparePassword(password);
+        if (!isMatch) return res.status(400).json({ message: 'Invalid email or password' });
+
+        const token = jwt.sign({ id: user._id }, process.env.JWT_SECRET, { expiresIn: '1h' });
+        res.json({ token });
+    } catch (error) {
+        console.error('Login error:', error);
+        res.status(500).json({ message: error.message });
+    }
+};
