+++ server/routes/auth.js
@@ -1,43 +1,11 @@
+
 const express = require('express');
 const router = express.Router();
-const User = require('../models/User');
-const jwt = require('jsonwebtoken');
-require('dotenv').config();
+const { registerUser, loginUser } = require('../controllers/authController');
 
-// Register Route
-router.post('/register', async (req, res) => {
-    const { username, email, password } = req.body;
-    try {
-        const userExists = await User.findOne({ email });
-        if (userExists) {
-            return res.status(400).json({ message: 'User already exists' });
-        }
+router.post('/register', registerUser);
 
-        const user = new User({ username, email, password });
-        await user.save();
-        res.status(201).send('User registered successfully');
-    } catch (error) {
-        res.status(500).json({ message: error.message });
-    }
-});
 
-// Login Route
-router.post('/login', async (req, res) => {
-    const { email, password } = req.body;
-    try {
-        const user = await User.findOne({ email });
-        if (!user) return res.status(400).json({ message: 'Invalid email or password' });
-
-        const isMatch = await user.comparePassword(password);
-        if (!isMatch) return res.status(400).json({ message: 'Invalid email or password' });
-
-        const token = jwt.sign({ id: user._id }, process.env.JWT_SECRET, { expiresIn: '1h' });
-        res.json({ token });
-    } catch (error) {
-        console.error('Login error:', error);
-        res.status(500).json({ message: error.message });
-    }
-});
-
+router.post('/login', loginUser);
 
 module.exports = router;
